Remind me to never try that again. Stats code will now update without redrawing all the pies, it will update in place and preserve all the filters that are applied.
