+++ pages/stats.html
@@ -49,7 +49,7 @@
       </td>
       <td>
         <div class="span4" id="dc-local-chart">
-          <h4>Top 10 Job Localities</h4>
+          <h4>Job Localities</h4>
         </div>
       </td>
       <td>
@@ -71,7 +71,7 @@
       </td>
       <td>
         <div class="span4" id="dc-propertytags-chart">
-          <h4>Job Location Tags</h4>
+          <h4>Property Tags</h4>
         </div>
       </td>
     </tr>
@@ -95,7 +95,7 @@
     <tr>
       <td colspan="3">
         <span>
-          <table class="table table-hover" id="dc-table-graph" style="width:1200px;margin-left:auto;margin-right:auto;align;text-align:left;">
+          <table class="table table-hover" id="dc-table-graph" style="width:90%;margin-left:auto;margin-right:auto;align;text-align:left;">
             <thead>
               <tr class="header">
                 <th>Job Number</th>
